trasladar el codigo a un solo componente
