+++ react-router/router-app/src/App.jsx
@@ -1,175 +1,181 @@
-/* eslint-disable react-refresh/only-export-components */
-/* eslint-disable react/no-unescaped-entities */
-/* eslint-disable react/prop-types */
-/* eslint-disable no-unused-vars */
+// /* eslint-disable react-refresh/only-export-components */
+// /* eslint-disable react/no-unescaped-entities */
+// /* eslint-disable react/prop-types */
+// /* eslint-disable no-unused-vars */
 
-import './App.css'
-import './index.css'
+// import './App.css'
+// import './index.css'
 
-import ReactDOM from 'react-dom/client'
-import { useState } from 'react'
+// import ReactDOM from 'react-dom/client'
+// import { useState } from 'react'
 
-import {
-  BrowserRouter as Router,
-  Routes,
-  Route,
-  Link,
-  Navigate,
-  useParams,
-  useNavigate,
-} from 'react-router-dom'
+// import {
+//   BrowserRouter as Router,
+//   Routes,
+//   Route,
+//   Link,
+//   Navigate,
+//   useParams,
+//   useNavigate,
+// } from 'react-router-dom'
 
-const Home = () => (
-  <div>
-    <h2>TKTL notes app</h2>
-    <p>
-      Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-      Lorem Ipsum has been the industry's standard dummy text ever since the
-      1500s, when an unknown printer took a galley of type and scrambled it to
-      make a type specimen book. It has survived not only five centuries, but
-      also the leap into electronic typesetting, remaining essentially
-      unchanged. It was popularised in the 1960s with the release of Letraset
-      sheets containing Lorem Ipsum passages, and more recently with desktop
-      publishing software like Aldus PageMaker including versions of Lorem
-      Ipsum.
-    </p>
-  </div>
-)
+// import React from 'react'
 
-const Note = ({ notes }) => {
-  const id = useParams().id
-  const note = notes.find((n) => n.id === Number(id))
-  return (
-    <div>
-      <h2>{note.content}</h2>
-      <div>{note.user}</div>
-      <div>
-        <strong>{note.important ? 'important' : ''}</strong>
-      </div>
-    </div>
-  )
-}
+// const Home = () => (
+//   <div>
+//     <h2>TKTL notes app</h2>
+//     <p>
+//       Lorem Ipsum is simply dummy text of the printing and typesetting industry.
+//       Lorem Ipsum has been the industry's standard dummy text ever since the
+//       1500s, when an unknown printer took a galley of type and scrambled it to
+//       make a type specimen book. It has survived not only five centuries, but
+//       also the leap into electronic typesetting, remaining essentially
+//       unchanged. It was popularised in the 1960s with the release of Letraset
+//       sheets containing Lorem Ipsum passages, and more recently with desktop
+//       publishing software like Aldus PageMaker including versions of Lorem
+//       Ipsum.
+//     </p>
+//   </div>
+// )
 
-const Notes = ({ notes }) => (
-  <div>
-    <h2>Notes</h2>
-    <ul>
-      {notes.map((note) => (
-        <li key={note.id}>
-          <Link to={`/notes/${note.id}`}>{note.content}</Link>
-        </li>
-      ))}
-    </ul>
-  </div>
-)
+// const Note = ({ notes }) => {
+//   const id = useParams().id
+//   const note = notes.find((n) => n.id === Number(id))
+//   return (
+//     <div>
+//       <h2>{note.content}</h2>
+//       <div>{note.user}</div>
+//       <div>
+//         <strong>{note.important ? 'important' : ''}</strong>
+//       </div>
+//     </div>
+//   )
+// }
 
-const Users = () => (
-  <div>
-    <h2>TKTL notes app</h2>
-    <ul>
-      <li>Matti Luukkainen</li>
-      <li>Juha Tauriainen</li>
-      <li>Arto Hellas</li>
-    </ul>
-  </div>
-)
+// const Notes = ({ notes }) => (
+//   <div>
+//     <h2>Notes</h2>
+//     <ul>
+//       {notes.map((note) => (
+//         <li key={note.id}>
+//           <Link to={`/notes/${note.id}`}>{note.content}</Link>
+//         </li>
+//       ))}
+//     </ul>
+//   </div>
+// )
 
-const Login = (props) => {
-  const navigate = useNavigate()
+// const Users = () => (
+//   <div>
+//     <h2>TKTL notes app</h2>
+//     <ul>
+//       <li>Matti Luukkainen</li>
+//       <li>Juha Tauriainen</li>
+//       <li>Arto Hellas</li>
+//     </ul>
+//   </div>
+// )
 
-  const onSubmit = (event) => {
-    event.preventDefault()
-    props.onLogin('mluukkai')
-    navigate('/')
-  }
+// const Login = (props) => {
+//   const navigate = useNavigate()
 
-  return (
-    <div>
-      <h2>login</h2>
-      <form onSubmit={onSubmit}>
-        <div>
-          username: <input />
-        </div>
-        <div>
-          password: <input type="password" />
-        </div>
-        <button type="submit">login</button>
-      </form>
-    </div>
-  )
-}
+//   const onSubmit = (event) => {
+//     event.preventDefault()
+//     props.onLogin('mluukkai')
+//     navigate('/')
+//   }
 
-const App = () => {
-  const [notes, setNotes] = useState([
-    {
-      id: 1,
-      content: 'HTML is easy',
-      important: true,
-      user: 'Matti Luukkainen',
-    },
-    {
-      id: 2,
-      content: 'Browser can execute only JavaScript',
-      important: false,
-      user: 'Matti Luukkainen',
-    },
-    {
-      id: 3,
-      content: 'Most important methods of HTTP-protocol are GET and POST',
-      important: true,
-      user: 'Arto Hellas',
-    },
-  ])
+//   return (
+//     <div>
+//       <h2>login</h2>
+//       <form onSubmit={onSubmit}>
+//         <div>
+//           username: <input />
+//         </div>
+//         <div>
+//           password: <input type="password" />
+//         </div>
+//         <button type="submit">login</button>
+//       </form>
+//     </div>
+//   )
+// }
 
-  const [user, setUser] = useState(null)
+// const App = () => {
+//   const [notes, setNotes] = useState([
+//     {
+//       id: 1,
+//       content: 'HTML is easy',
+//       important: true,
+//       user: 'Matti Luukkainen',
+//     },
+//     {
+//       id: 2,
+//       content: 'Browser can execute only JavaScript',
+//       important: false,
+//       user: 'Matti Luukkainen',
+//     },
+//     {
+//       id: 3,
+//       content: 'Most important methods of HTTP-protocol are GET and POST',
+//       important: true,
+//       user: 'Arto Hellas',
+//     },
+//   ])
 
-  const login = (user) => {
-    setUser(user)
-  }
+//   const [user, setUser] = useState(null)
 
-  const padding = {
-    padding: 5,
-  }
+//   const login = (user) => {
+//     setUser(user)
+//   }
 
-  return (
-    <div>
-      <Router>
-        <div>
-          <Link style={padding} to="/">
-            home
-          </Link>
-          <Link style={padding} to="/notes">
-            notes
-          </Link>
-          <Link style={padding} to="/users">
-            users
-          </Link>
-          {user ? (
-            <em>{user} logged in</em>
-          ) : (
-            <Link style={padding} to="/login">
-              login
-            </Link>
-          )}
-        </div>
+//   const padding = {
+//     padding: 5,
+//   }
 
-        <Routes>
-          <Route path="/notes/:id" element={<Note notes={notes} />} />
-          <Route path="/notes" element={<Notes notes={notes} />} />
-          <Route
-            path="/users"
-            element={user ? <Users /> : <Navigate replace to="/login" />}
-          />
-          <Route path="/login" element={<Login onLogin={login} />} />
-          <Route path="/" element={<Home />} />
-        </Routes>
-      </Router>
-      <div>
-        <br />
-        <em>Note app, Department of Computer Science 2023</em>
-      </div>
-    </div>
-  )
-}
+//   return (
+//     <div>
+//       <Router>
+//         <div>
+//           <Link style={padding} to="/">
+//             home
+//           </Link>
+//           <Link style={padding} to="/notes">
+//             notes
+//           </Link>
+//           <Link style={padding} to="/users">
+//             users
+//           </Link>
+//           {user ? (
+//             <em>{user} logged in</em>
+//           ) : (
+//             <Link style={padding} to="/login">
+//               login
+//             </Link>
+//           )}
+//         </div>
 
-export default App
+//         <Routes>
+//           <Route path="/notes/:id" element={<Note notes={notes} />} />
+//           <Route path="/notes" element={<Notes notes={notes} />} />
+//           <Route
+//             path="/users"
+//             element={user ? <Users /> : <Navigate replace to="/login" />}
+//           />
+//           <Route path="/login" element={<Login onLogin={login} />} />
+//           <Route path="/" element={<Home />} />
+//         </Routes>
+//       </Router>
+//       <div>
+//         <br />
+//         <em>Note app, Department of Computer Science 2023</em>
+//       </div>
+//     </div>
+//   )
+// }
+
+// ReactDOM.createRoot(document.getElementById('root')).render(
+//   <React.StrictMode>
+//     <App />
+//   </React.StrictMode>
+// )
